Move compare.go into util package
